Page info adjustments
Changes to Home Page with inclusion of body.
Changes to Home Page with the link button
Changes to CSS document with reference to positioning on content in body.
Changes to About Me page with reference to Hobbies and experience.
Changes to Work Page with Reference to Positioning of content in Tables.